Refactor EntryLogger class
- Split EntryLogger class into multiple classes
- create separate classes for EntryLogManager, EntryLoggerAllocator and EntryLogManagerForSingleEntryLog
